+++ src/main/java/home/des/example/lesson2/ReadFiles.java
@@ -1,85 +1,79 @@
 package home.des.example.lesson2;
 
-import java.io.File;
-import java.io.IOException;
+import org.postgresql.shaded.com.ongres.scram.common.bouncycastle.pbkdf2.Strings;
+
+import java.io.*;
 import java.nio.file.Files;
-import java.nio.file.LinkOption;
 import java.nio.file.Path;
 import java.nio.file.Paths;
+import java.util.ArrayList;
+import java.util.Arrays;
 
 public class ReadFiles {
-
 }
 
-class Read1 {
-    public static void main(String[] args) {
-        Path pathFile = Paths.get("newFile.txt");
-        System.out.println("Print info of file");
-        System.out.println("\t file name: " + pathFile.getFileName());
-        System.out.println("\t root of the path: " + pathFile.getRoot());
-        System.out.println("\t parent of the target: " + pathFile.getParent());
+class ReadOneFileByString {
+    public static void main(String[] args) throws IOException {
+        Path pathFile = Paths.get("data\\fiftyBytes.txt");
+        File file = new File(String.valueOf(pathFile));
 
-        for (Path element : pathFile) {
-            System.out.println("\t path element: " + element);
+        BufferedReader reader = new BufferedReader(new FileReader(file));
+        String tmp = reader.readLine();
+
+        while (tmp != null) {
+            System.out.println(tmp);
+            tmp = reader.readLine();
         }
     }
 }
 
-class Read2 {
+class ReadOneFileByBytes {
     public static void main(String[] args) throws IOException {
-        Path pathFile = Paths.get("newFile.txt");
-        Path testPathNormalized = Paths
-                .get(pathFile.normalize().toString());
-        System.out.println("It's normalized absolute path is: "
-                + testPathNormalized.toAbsolutePath());
-        System.out.println("It's normalized real path is: "
-                + pathFile.toRealPath(LinkOption.NOFOLLOW_LINKS));
+        Path pathFile = Paths.get("data/fiftyBytes.txt");
+        File file = new File(String.valueOf(pathFile));
+
+        FileInputStream fis = new FileInputStream(file);
+
+        byte[] arr = Files.readAllBytes(pathFile);
+
+        System.out.println(Strings.asCharArray(arr));
     }
 }
 
-class Read3 {
+class UniteFiveFiles {
     public static void main(String[] args) throws IOException {
-        Path pathFile = Paths.get("newFile.txt");
-        Files.createFile(pathFile);
-        Path oldFile = Paths.get("oldFile.txt");
-        Files.createFile(oldFile);
-        if (Files.exists(oldFile)) {
-            System.out.println("The file/directory " + oldFile.getFileName() + " exists");
-            if (Files.isDirectory(oldFile)) {
-                System.out.println("The " + oldFile.getFileName() + " is directory");
-            } else {
-                System.out.println("The " + oldFile.getFileName() + " is file");
+        String allPath = "data/";
+        ArrayList<InputStream> inputStreams = new ArrayList<>();
+        inputStreams.add(new FileInputStream(allPath + "fiftyBytes.txt"));
+        for (int i = 1; i < 5; i++) {
+            inputStreams.add(new FileInputStream(allPath + "fiftyBytes" + i + ".txt"));
+        }
+        byte[] arr = new byte[4096];
+        int count;
+        for (InputStream is : inputStreams) {
+            while ((count = is.read(arr)) != -1) {
+                System.out.print(new String(arr, 0, count));
             }
+            System.out.println();
         }
-
-        Files.delete(pathFile);
-        Files.delete(oldFile);
     }
 }
 
-class Read4 {
+class ReadFileByPages {
     public static void main(String[] args) throws IOException {
-        Path pathFile = Paths.get("newFile.txt");
-        Files.createFile(pathFile);
-        Path oldFile = Paths.get("oldFile.txt");
-        try {
-            Files.copy(pathFile, oldFile);
-        } catch (IOException e) {
-            e.printStackTrace();
+        String allPath = "data/";
+        InputStream is = new FileInputStream(allPath + "fiftyBytes.txt");
+        byte[] arr = new byte[1800];
+        int count;
+        int page = 1;
+        while ((count = is.read(arr)) != -1) {
+            System.out.println("Page: " + page + "\n");
+            System.out.print(new String(arr, 0, count));
+            System.out.print("\n\n");
+            page++;
         }
+        System.out.println();
     }
 }
 
-class Read5{
-    public static void main(String[] args) throws IOException {
-        Path pathFile = Paths.get("newFile.txt");
-        Files.createFile(pathFile);
-        Path oldFile = Paths.get("oldFile.txt");
-        if (!Files.exists(oldFile)){
-            File file = new File(String.valueOf(pathFile));
-            file.renameTo(new File(String.valueOf(oldFile)));
-        } else {
-            System.out.println("File " + oldFile.toAbsolutePath() + " exists");
-        }
-    }
-}
+
